philo/src/routine.c: changed data type of argument 'start_time' from 'int' to 'long'

+++ philo/philosophers.h
@@ -6,7 +6,7 @@
 /*   By: we <we@student.42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/28 14:33:55 by tjun-yu           #+#    #+#             */
-/*   Updated: 2024/05/31 17:38:08 by we               ###   ########.fr       */
+/*   Updated: 2024/05/31 17:57:25 by we               ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -54,11 +54,11 @@ typedef struct s_table
 
 void	start_simulation(t_table *table);
 void	*philo_routine(void *arg);
-void	p_eat(t_philo *philo, int time_to_eat, int start_time);
-void	p_take_fork(t_philo *philo, int start_time);
-void	p_sleep(t_philo *philo, int time_to_sleep, int start_time);
-void	p_think(t_philo *philo, int start_time);
-void	p_die(t_philo *philo, int start_time);
+void	p_eat(t_philo *philo, int time_to_eat, long start_time);
+void	p_take_fork(t_philo *philo, long start_time);
+void	p_sleep(t_philo *philo, int time_to_sleep, long start_time);
+void	p_think(t_philo *philo, long start_time);
+void	p_die(t_philo *philo, long start_time);
 
 void	init_env(t_table *table, char *argv[]);
 t_philo	*init_philos(char *forks, int count);
